Moved AccessibilityService from /shared to /core.
