+++ src/app/core/services/accessibility.service.ts
